Tell Apollo when HLS manifest is ready and fallback to software decode
When testing stuff I realized that some video files
are having issues which seem related to the `uridecodebin`
and the NVIDIA VP9 hardware decoder.

The target duration is completely messed up in these cases
and causes some  troubles in playback while the transcode is still running.

If this specific malfunction is detected in the HLS manifest,
the pipeline is shut down and a new one is created which
is configured to only use software decoders.

I've created an issue at gstreamer and hopefully
this can be fixed in the future:
https://gitlab.freedesktop.org/gstreamer/gst-plugins-bad/-/issues/1753